Refactor: Flatten module structure and remove dead code
- Deleted dead code `src/cli.rs`.
- Flattened `src/ast/` into `src/ast.rs`.
- Flattened `src/codegen/` into `src/codegen.rs` (merged `types.rs`, `utils.rs`, `rust.rs`).
- Flattened `src/parser/` logic into `src/parser.rs` (merged `mod.rs`, `builder.rs`).
- Flattened `src/grammar/` into `src/grammar.rs` and moved pest file to `src/grammar.pest`.
- Updated `tests/option_result_tests.rs` import path.
- Updated `.jules/razor.md` with reduction log.

This adheres to the Razor philosophy of flattening unnecessary hierarchies and deleting unused code.
Verified with `cargo check`, `cargo test`, `cargo clippy`, and `cargo fmt`.
