+++ src/ast.rs
@@ -1,7 +1,40 @@
-//! AST node definitions for ΓΛΩΣΣΑ
+//! Abstract Syntax Tree for ΓΛΩΣΣΑ
 //!
-//! These nodes capture the structure of a GLOSSA program,
-//! preserving both original Greek text and normalized forms.
+//! The AST captures the semantic structure of a GLOSSA program,
+//! preserving morphological information from Greek words.
+//!
+//! # The Tree Structure
+//!
+//! The AST hierarchy reflects the grammatical structure of the language:
+//!
+//! * [`Program`]: The root node, containing a list of statements.
+//! * [`Statement`]: A sentence, ending with a period (`.`) or query mark (`?` / `;`).
+//!   * A statement consists of one or more [`Clause`]s.
+//! * [`Clause`]: A comma-separated part of a statement.
+//!   * Example: `ὁ ἄνθρωπος, τὸν λόγον λέγει.` (Two clauses: "The man", "says the word").
+//! * [`Expr`]: An expression (word, literal, operation).
+//!   * [`Expr::Word`]: A raw Greek word with its original and normalized forms.
+//!
+//! # Design Philosophy
+//!
+//! Unlike traditional ASTs that might discard surface-level details, the GLOSSA AST
+//! preserves the *original* Greek text in [`Word`] nodes. This is crucial for:
+//!
+//! 1. **Error Reporting**: Using the original polytonic Greek in error messages.
+//! 2. **Morphological Analysis**: The semantic phase needs the original form to
+//!    distinguish subtle variations if needed.
+//!
+//! # Example
+//!
+//! A simple program like `«χαῖρε» λέγε.` produces:
+//!
+//! ```text
+//! Program
+//! └── Statement::Regular
+//!     └── Clause
+//!         ├── Expr::StringLiteral("χαῖρε")
+//!         └── Expr::Word("λέγε")
+//! ```
 
 use smol_str::SmolStr;
