--- src/codegen/mod.rs
@@ -1,99 +1,0 @@
-//! Code generation for ΓΛΩΣΣΑ
-//!
-//! This module handles the translation of the Analyzed Program into executable Rust code.
-//!
-//! # The Strategy: Transpilation
-//!
-//! ΓΛΩΣΣΑ compiles to Rust rather than LLVM IR or Assembly for several strategic reasons.
-//! This approach is often called "transpilation" or "source-to-source compilation".
-//!
-//! ## 1. Safety First (The "Rust" Shield)
-//! By targeting Rust, we inherit the most sophisticated borrow checker and memory safety
-//! system in the world. If the ΓΛΩΣΣΑ compiler generates valid Rust code, the resulting
-//! binary is guaranteed to be memory-safe (no segfaults, no data races).
-//!
-//! ## 2. The Ecosystem
-//! Users get instant access to `cargo`, `crates.io`, and the entire Rust ecosystem.
-//! A ΓΛΩΣΣΑ program is just a Rust program with a Greek syntax.
-//!
-//! ## 3. Performance
-//! `rustc` is an incredibly sophisticated optimizing compiler. We don't need to write
-//! our own optimization passes (constant folding, loop unrolling, etc.) because `rustc`
-//! does it for us.
-//!
-//! # The Mapping Strategy
-//!
-//! The mapping between ΓΛΩΣΣΑ and Rust is direct and intentional.
-//!
-//! | ΓΛΩΣΣΑ Concept | Rust Equivalent | Notes |
-//! |----------------|-----------------|-------|
-//! | **Types** | | |
-//! | `ἀριθμός` (Number) | `i64` | 64-bit signed integer |
-//! | `ὄνομα` (String) | `String` | Heap-allocated UTF-8 string |
-//! | `ἀληθές/ψεῦδος` | `bool` | Boolean |
-//! | `λίστη` | `Vec<T>` | Dynamic array |
-//! | `σύνολον` | `HashSet<T>` | Unique collection |
-//! | `χάρτης` | `HashMap<K, V>` | Key-value store |
-//! | **Control Flow** | | |
-//! | `εἰ ...` (If) | `if ...` | Standard control flow |
-//! | `εἰ δὲ μή ...` | `else ...` | Else block |
-//! | `ἕως ...` (While) | `while ...` | While loop |
-//! | `διὰ ...` (For) | `for ... in ...` | Iterator loop |
-//! | `παῦε` (Break) | `break` | Break loop |
-//! | `συνέχιζε` (Continue)| `continue` | Continue loop |
-//! | **Structures** | | |
-//! | `εἶδος` (Struct) | `struct` | Generates `#[derive(Clone, Debug, ...)]` |
-//! | `χαρακτήρ` (Trait) | `trait` | Maps methods directly |
-//! | `ὁρίζειν` (Define) | `impl` | Implementation block |
-//! | **Error Handling** | | |
-//! | `ἀποτέλεσμα` | `Result<T, E>` | Uses standard Rust Result |
-//! | `εὑρεθείη` | `Option<T>` | The "Optative" mood |
-//! | `;` (Propagate) | `?` | The Try operator |
-//! | `!` (Unwrap) | `.unwrap()` | Panic on failure |
-//!
-//! # Example Translation
-//!
-//! **ΓΛΩΣΣΑ Source:**
-//! ```text
-//! // Define a struct
-//! εἶδος Point ὁρίζειν {
-//!     x ἀριθμοῦ.
-//!     y ἀριθμοῦ.
-//! }.
-//!
-//! // Check equality
-//! ξ πέντε ἔστω.
-//! εἰ ξ πέντε ἰσοῦται,
-//!     «ἰσχύει» λέγε.
-//! ```
-//!
-//! **Generated Rust:**
-//! ```rust,ignore
-//! #[derive(Clone, Debug, PartialEq)]
-//! struct Point {
-//!     x: i64,
-//!     y: i64,
-//! }
-//!
-//! fn main() {
-//!     let x = 5;
-//!     if x == 5 {
-//!         println!("ἰσχύει");
-//!     }
-//! }
-//! ```
-//!
-//! # Architecture
-//!
-//! The codegen process uses the [`quote`] crate to construct a Rust AST (TokenStream)
-//! directly from the Semantic Analysis model. This ensures that the generated code
-//! is syntactically valid Rust and avoids "stringly typed" code generation errors.
-//!
-//! 1. **Semantic Analysis**: Produces an [`AnalyzedProgram`](crate::semantic::AnalyzedProgram).
-//! 2. **Code Generation**: [`generate_rust`] takes the program and produces a Rust source string.
-
-mod rust;
-pub mod types;
-pub(crate) mod utils;
-
-pub use rust::*;
